pyproject.toml
[project]
name = "homepage"
version = "1.0.0"
description = "Personal homepage built with Django"
requires-python = ">=3.12"
dependencies = [
    "django>=6.0.2",
    "gunicorn==25.1.0",
    "psycopg2-binary==2.9.11",
    "python-decouple>=3.8",
    "markdown>=3.10.2",
    "sentry-sdk>=2.53.0",
    "django-debug-toolbar>=6.2.0",
    "httpx>=0.28.1",
    "djangorestframework>=3.16.1",
    "drf-spectacular>=0.29.0",
    "psutil>=7.2.2",
    "whitenoise>=6.12.0",
    "django-health-check>=4.0.6",
    "ruff>=0.15.4",
    "mypy>=1.19.1",
    "django-stubs>=5.2.9",
    "djangorestframework-stubs>=3.16.8",
    "pre-commit>=4.5.1",
    "types-markdown>=3.10.2.20260211",
    "django-htmx>=1.27.0",

]

[tool.setuptools]
packages = []

[tool.ruff]
line-length = 120
target-version = "py312"

[tool.ruff.lint]
select = [
    "E",      # pycodestyle errors
    "W",      # pycodestyle warnings
    "F",      # Pyflakes
    "I",      # isort
    "B",      # flake8-bugbear
    "C4",     # flake8-comprehensions
    "UP",     # pyupgrade
    "DJ",     # flake8-django
]
ignore = [
    "E501",   # line too long (handled by formatter)
]

[tool.ruff.lint.isort]
known-first-party = ["core", "blog", "showcase", "config"]

[tool.mypy]
python_version = "3.14"
plugins = ["mypy_django_plugin.main", "mypy_drf_plugin.main"]
strict = false
warn_return_any = true
warn_unused_ignores = true
ignore_missing_imports = true
exclude = ["migrations", "venv", ".venv"]

[tool.django-stubs]
django_settings_module = "config.settings.development"


